Remove unused GHA actions
Note that this was found due to a security assessment tracing down
uses of `tj-actions/changed-files` github actions sources. Github CI
is not used on this repo and these files are located in a `.github-old`
(versus a `.github) so they wouldn't have been picked up anyhow.

For the sake of making sure no one copies and pastes them willy nilly
back in. Lets go ahead and remove them.